requirements.txt
cloudpickle==1.2.1
cmake==3.13.3
future==0.17.1
gym==0.13.0
gym-retro==0.7.0
gym-super-mario-bros==7.2.3
neat-python==0.92
nes-py==8.0.2
numpy==1.16.4
opencv-python==4.1.0.25
pkg-resources==0.0.0
pyglet==1.4.0
scipy==1.3.0
six==1.12.0
tqdm==4.32.2
